apps/tool-router/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-gateway"
version = "0.1.1"
description = "Self-hosted MCP gateway (Context Forge) with tool-router"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0,<2.0.0",
    "fastmcp>=0.1.0,<1.0.0",
    "requests>=2.31.0,<3.0.0",
    "flask>=3.0.0,<4.0.0",
    "pyyaml>=6.0.0,<7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests", "src/tool_router/ai/tests", "src/tool_router/api/tests", "src/tool_router/core/tests"]
pythonpath = ["src"]
addopts = "-v --strict-markers"

[tool.coverage.run]
source = ["src/tool_router"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "src/tool_router/__main__.py",
    "src/tool_router/core/server.py",
    "src/tool_router/tools/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2

[tool.ruff]
line-length = 120
target-version = "py311"
src = ["tool_router"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PD",     # pandas-vet
    "PGH",    # pygrep-hooks
    "PL",     # Pylint
    "TRY",    # tryceratops
    "NPY",    # NumPy-specific rules
    "RUF",    # Ruff-specific rules
]

# Ignore specific rules that may be too strict or conflict with project style
ignore = [
    "ANN401",  # Dynamically typed expressions (Any) are disallowed
    "S101",    # Use of assert detected (pytest uses assert)
    "PLR2004", # Magic value used in comparison
    "S310",    # Audit URL open (urllib is used intentionally for HTTP requests)
    "B904",    # raise from err (not always needed for simple error wrapping)
    "G004",    # Logging f-strings (acceptable for structured logging)
    "TRY401",  # Redundant exception in logging.exception (intentional for clarity)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]  # Allow unused imports in __init__.py
"tool_router/tests/**/*.py" = [
    "S101",    # Allow assert in tests
    "ANN",     # Don't require type hints in tests
    "PLR2004", # Allow magic values in tests
    "ARG",     # Allow unused arguments in test fixtures
    "SIM117",  # Allow nested with statements in tests
    "PT011",   # Allow broad exception types in pytest.raises
    "B017",    # Allow asserting on Exception in tests
    "PLC0415", # Allow imports inside test functions
]
"tool_router/test_*.py" = [
    "S101",    # Allow assert in tests
    "ANN",     # Don't require type hints in tests
    "SIM117",  # Allow nested with statements in tests
]
"tool_router/ai/tests/**/*.py" = [
    "S101",    # Allow assert in tests
    "ANN",     # Don't require type hints in tests
    "PLR2004", # Allow magic values in tests
    "ARG",     # Allow unused arguments in test fixtures
    "SIM117",  # Allow nested with statements in tests
    "PT011",   # Allow broad exception types in pytest.raises
    "B017",    # Allow asserting on Exception in tests
    "PLC0415", # Allow imports inside test functions
]
"tool_router/api/tests/test_ide_config.py" = [
    "S101",    # Allow assert statements in tests
    "S106",    # Allow hardcoded test tokens/passwords in tests
    "PLR2004", # Allow magic values in tests
]
"tool_router/api/tests/test_rest.py" = [
    "TRY003",  # Allow long exception messages for validation
    "EM101",   # Allow string literals in exceptions for validation
    "EM102",   # Allow f-strings in exceptions for validation
]
"tool_router/api/ide_config.py" = [
    "TRY003",  # Allow long exception messages for validation
    "EM101",   # Allow string literals in exceptions for validation
    "EM102",   # Allow f-strings in exceptions for validation
]
"tool_router/api/rest.py" = [
    "TRY300",  # Allow return in try block for REST handlers
    "BLE001",  # Allow catching Exception for REST error handling
    "ANN202",  # Allow missing return annotations in nested functions
    "T201",    # Allow print in __main__ for standalone server
    "S104",    # Allow 0.0.0.0 binding for development server
    "B008",    # Allow Body(...) in FastAPI defaults
]
"tool_router/api/tests/**/*.py" = [
    "S101",    # Allow assert in tests
    "ANN",     # Don't require type hints in tests
    "PLR2004", # Allow magic values in tests
    "ARG",     # Allow unused arguments in test fixtures
]
"tool_router/server.py" = [
    "PLR0911", # Allow more return statements in main routing function
    "E501",    # Allow long lines in docstrings
]
"tool_router/core/server.py" = [
    "PLR0911", # Allow more return statements in main routing function
    "PLR0912", # Allow more branches in execute_task (AI + keyword logic)
    "PLR0915", # Allow more statements in execute_task
    "PLW0603", # Allow global statement for config/ai_selector initialization
    "BLE001",  # Allow broad exception catching for fallback logic
    "E501",    # Allow long lines in docstrings
]
"tool_router/gateway/client.py" = [
    "PLC0415", # Allow conditional imports for backward compatibility
    "E501",    # Allow long lines in list comprehensions
]
"tool_router/gateway_client.py" = [
    "PLC0415", # Allow conditional imports
]
"tool_router/scoring/matcher.py" = [
    "E501",    # Allow long lines in scoring calculations
    "PLR0913", # Allow more arguments in hybrid scoring function
]
"tool_router/ai/selector.py" = [
    "PLR0911", # Allow more return statements in parser logic
    "TRY300",  # Allow direct returns in try blocks for readability
    "ANN204",  # Skip __init__ return type annotation
]

[tool.ruff.lint.isort]
known-first-party = ["tool_router"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 5
max-branches = 12
max-returns = 6
max-statements = 50

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
